+++ database/migrations/2021_12_16_185816_add_div_classifier_to_paneles_table.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddDivClassifierToPanelesTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('paneles', function (Blueprint $table) {
+            $table->unsignedBigInteger('div_classifier')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('paneles', function (Blueprint $table) {
+            $table->dropColumn('div_classifier');
+        });
+    }
+}